Minor changes in file names
