+++ client/package.json
@@ -7,6 +7,7 @@
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@mui/icons-material": "^5.11.0",
+    "@mui/lab": "^5.0.0-alpha.117",
     "@mui/material": "^5.11.6",
     "@mui/styled-engine-sc": "^5.11.0",
     "@react-oauth/google": "^0.6.1",
@@ -16,6 +17,7 @@
     "axios": "^1.2.4",
     "jwt-decode": "^3.1.2",
     "moment": "^2.29.4",
+    "mui-chips-input": "^1.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-file-base64": "^1.0.3",
