+++ client/src/components/PostDetails/PostDetails.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const PostDetails = () => {
+    console.log("POST_DETAILS")
+    return (
+        <div>
+            
+        </div>
+    );
+};
+
+export default PostDetails;
