+++ client/src/components/styles.css
@@ -1,0 +1,3 @@
+.ul{
+    justify-content: space-around;
+}
